backend/pyproject.toml
[project]
name = "LogosAI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.19.0",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.34.3",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"llm/prompts.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.11.0",
]
